requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2024.12.14
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
dateparser==1.2.0
Django>=5.1,<5.2
django-allauth==65.4.1
django-extensions==3.2.3
django-htmx==1.22.0
django-split-settings==1.3.2
django-timescaledb
dj-database-url
django-celery-beat 
django-celery-results
django-admin-interface
django-admin-rangefilter
fonttools==4.55.8
frozendict==2.4.6
frozenlist==1.5.0
html5lib==1.1
idna==3.10
kiwisolver==1.4.8
llvmlite==0.44.0
lxml==5.3.0
matplotlib==3.10.0
multidict==6.1.0
multitasking==0.0.11
numba==0.61.0
numpy==1.26.4
packaging==24.2
pandas==2.2.3
pandas-datareader==0.10.0
pandas_ta @ git+https://github.com/twopirllc/pandas-ta.git@438f3a97b71b26bc17efb1a204fb0e0fdc48e0aa
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
pycryptodome==3.21.0
pyparsing==3.2.1
python-binance==1.0.27
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
ta==0.11.0
typing_extensions==4.12.2
tzdata==2025.1
tzlocal==5.2
urllib3==2.3.0
webencodings==0.5.1
websockets==14.2
yarl==1.18.3
yfinance==0.2.52
jupyter
psycopg[binary]
Celery
Redis
openai